# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS API
  version: 1.0.0
  description: |
    A RESTful Learning Management System API built with Node.js, Express.js,
    MongoDB, and Mongoose.

    The API provides authentication, user management, course management,
    lessons, enrollments, progress tracking, reviews, favorites,
    instructor dashboard functionality, and Stripe payments.
servers:
  - url: http://localhost:4000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      description: Course ID
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
        example: 65f1a2b3c4d5e6f789012345
...
    UserId:
      name: userId
      in: path
      required: true
      description: User ID
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
        example: 65f1a2b3c4d5e6f789012347
    ReviewId:
      name: reviewId
      in: path
      required: true
      description: Review ID
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
        example: 65f1a2b3c4d5e6f789012348
...
  schemas:
...
    UserSummary:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012345
        firstName:
          type: string
          example: Ahmed
        lastName:
          type: string
          example: Mohamed
        email:
          type: string
          format: email
          example: ahmed@example.com
...
    UserByIdResponse:
      type: object
      properties:
        userId:
          type: string
          example: 65f1a2b3c4d5e6f789012345
        firstName:
          type: string
          example: Ahmed
        lastName:
          type: string
          example: Mohamed
        email:
          type: string
          format: email
          example: ahmed@example.com
...
    Review:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012352
        user:
          type: string
          example: 65f1a2b3c4d5e6f789012345
        course:
          type: string
          example: 65f1a2b3c4d5e6f789012346
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: Excellent course!
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReviewPopulated:
      type: object
      properties:
        _id:
          type: string
          example: 65f1a2b3c4d5e6f789012352
        user:
          $ref: '#/components/schemas/UserSummary'
        course:
          type: string
          example: 65f1a2b3c4d5e6f789012346
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: Excellent course!
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateReviewRequest:
      type: object
      required:
        - courseId
        - rating
      properties:
        courseId:
          type: string
          example: 65f1a2b3c4d5e6f789012346
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: Excellent course!
    UpdateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        comment:
          type: string
          example: Very good course.
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/ReviewPopulated'
...
    ReviewResponse:
      type: object
      properties:
        review:
          $ref: '#/components/schemas/Review'
    DeleteReviewResponse:
      type: object
      properties:
        message:
          type: string
          example: Review deleted successfully
        review:
          $ref: '#/components/schemas/Review'
...
    Error:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Something went wrong
paths:
...
  /api/users/{userId}:
    get:
      tags:
        - Users
      operationId: getUserById
      summary: Get user by ID
      description: Returns a user by ID. Administrator access is required.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserByIdResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/reviews/create:
    post:
      tags:
        - Reviews
      operationId: createReview
      summary: Create a course review
      description: |
        Creates a review for a published course.

        The authenticated user must be successfully enrolled in the course
        and must not already have a review for it.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User is not allowed to review this course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User has already reviewed this course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reviews/course/{courseId}:
    get:
      tags:
        - Reviews
      operationId: getCourseReviews
      summary: Get reviews for a course
      description: Returns all reviews for the specified course.
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          description: Course reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/reviews/{reviewId}:
    put:
      tags:
        - Reviews
      operationId: updateReview
      summary: Update a review
      description: Only the owner of the review can update it.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User is not the owner of this review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Reviews
      operationId: deleteReview
      summary: Delete a review
      description: Only the owner of the review can delete it.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '200':
          description: Review deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteReviewResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: User is not the owner of this review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...